Loadout System, HUD Timer, MapVote
A looooot of changes with this update. Implemented the loadout system (currently only through command, however, no menu) and managed to get the game timer displaying properly in the upper right corner of the HUD. Added a map vote at the end of round that REQUIRES separate MapVote addon. Fixed a few bugs and changed the player class to work with the new loadout system.