Major UI Update
Added a user menu to select loadout. Currently requires "M9K Specialties" installed and "lostcoast" (from L4D2) installed and mounted in order to test the model display in the menu. The model display does not work with the default weapons. Valid weapons, as well as a list of ammo to be given, has all been shifted to its own file.